still working in it

+++ ft_printvoid.c
@@ -5,20 +5,20 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mxaba <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/07/19 09:13:13 by mxaba             #+#    #+#             */
-/*   Updated: 2018/07/19 09:16:30 by mxaba            ###   ########.fr       */
+/*   Created: 2018/07/23 12:03:49 by mxaba             #+#    #+#             */
+/*   Updated: 2018/07/23 12:03:51 by mxaba            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "printf.h"
+#include "libprt.h"
 
 int		ft_printvoid(va_list ap)
 {
 	char			*p;
-	unsigned int	c;
+	unsigned long	c;
 
 	ft_putstr("0x");
-	c = va_arg(ap, unsigned int);
+	c = va_arg(ap, unsigned long);
 	p = ft_htoa(c);
 	ft_putstr(p);
 	return (ft_strlen(p) + 2);
